Add support for running Podman in the sbatch scripts
Add support for adding container images to the $installation_files_dir.
These container images can after the installation be used by
Podman when it is run in Sbatch scripts.

To make this work Podman needs to be run from within Podman. See also
https://stackoverflow.com/questions/64509618/podman-in-podman-similar-to-docker-in-docker

At the moment the code is more of a proof-of-concept.
Add the file desing.md to describe the design a bit more.
It is likely that the portmapping and files under adjusting_ports_for_norouter/
could be simplified and maybe even removed.

The base image of the Dockerfile has been changed to
quay.io/podman/stable:v2.1.1
that is based on Fedora 33

Signed-off-by: Erik Sjölund <erik.sjolund@gmail.com>